Fix prereq load bug; adopt higher-level syntax
We weren't reading the list of requires from the module description code
correctly.  As a result, modules such as APROPOS were not having their
prereqs loaded, and things weren't working.

We've also added a thin syntax layer to avoid quoting module names.
This change may not stick; we're not entirely sure it is necessary.