Support top-level `LET` expressions
The following code from `yasos.scm` was breaking module loading.  It
wants to use `let` as it does to create an `EQ?`-unique object instance.

    (let ((instance-tag "instance") ) ;; Make a unique tag within a local scope.
      ;; No other data object is EQ? to this tag.
      (set! make-instance
            (lambda (dispatcher) (cons instance-tag dispatcher)))
      (set! instance?
            (lambda (obj) (and (pair? obj) (eq? (car obj) instance-tag)))))

The solution in LOAD-MODULE was to inspect the symbol in the "operator"
position, and return `#f` if the operator is not one of the defined
operators -- for now `DEFINE` and `DEFINE-SYNTAX` (see the diff for
details).