Internal DEFINEs of lists are now allowed
E.g., you can now write inside a module code like

    (define my-list '(1 2 3))

Previously, `LOAD-MODULE` was blowing up when trying to load code like
this.  For example, this line from SRFI-25 would explode:

    (define (array:shape-index) '#(2 1 0))

This was happening because `MAP*` did not know about vectors, so it was
calling `CAR` on them.

This update fixes the problem by telling `MAP*` to basically ignore
vectors and keep walking down the tree.

Note: I'm not 100% convinced this is the right thing to do.  In the
future, we may want to change this so that `MAP*` calls its `LEAF-FUNC`
argument on the vector instead of ignoring it.  However, this patch
seems to work for now.